RF: Address review feedback on zarr-python 2/3 compat work
Code review (eba48849..5f87b6a8) flagged a few rough edges. None
change behaviour:

* Move `zarr_format_of` and the `TWO_ARRAY_ZARR_LAYOUT` mapping from
  `dandi/tests/test_files.py` to `dandi/tests/test_helpers.py` and drop
  the leading underscore from the now genuinely cross-module name. The
  previous arrangement cross-imported between test modules (including
  across packages from `dandi/cli/tests/test_digest.py`), which would
  silently break on rename or pytest collection-order changes. The
  underscore-prefixed name was also misleading since the symbol was
  imported elsewhere. `test_helpers.py` is the existing pattern for
  shared test utilities — keeping the `test_*.py` filename so pytest
  preserves assertion rewriting on the `assert` inside `zarr_format_of`.

* Collapse the two byte-identical `ValidationResult` arms in
  `ZarrAsset.get_validation_errors` into one append, with the message
  chosen via a conditional. The two arms differed only in the message
  string and risked drift over time.

* Add a `lgr.debug` line in `_is_empty_group`'s broad-`except` so that
  a real bug (as opposed to the expected zarr-python 3.x eager-iteration
  raise) is at least discoverable in debug logs rather than silently
  swallowed.

* Tighten `test_download_different_zarr_delete_dir` to assert the V3
  invariant (`initial_subdirs == {"c"}`) instead of just skipping the
  no-subdirs check on V3. The test still exercises the contrast
  between the smaller single-array Zarr and the bigger two-array one
  that the download must reconcile.

Verified mypy clean against zarr 3.1.5; pytest passes against zarr
2.18.7 and 3.1.5; pre-commit hooks all pass.
