+++ dandi/tests/test_download.py
@@ -27,8 +27,7 @@ import zarr
 
 from .fixtures import SampleDandiset, SampleDandisetFactory
 from .skip import mark
-from .test_files import _TWO_ARRAY_ZARR_LAYOUT, zarr_format_of
-from .test_helpers import assert_dirtrees_eq
+from .test_helpers import TWO_ARRAY_ZARR_LAYOUT, assert_dirtrees_eq, zarr_format_of
 from ..consts import DRAFT, dandiset_metadata_file
 from ..dandiarchive import DandisetURL
 from ..download import (
@@ -464,9 +463,7 @@ def test_download_different_zarr_onto_excluded_dotfiles(
     # `zarr_dandiset` was uploaded with default ``zarr.save`` — its on-disk
     # layout (and therefore the layout we expect to see after download) is
     # V2 for zarr-python 2.x, V3 for zarr-python 3.x.
-    layout = _TWO_ARRAY_ZARR_LAYOUT[
-        zarr_format_of(zarr_dandiset.dspath / "sample.zarr")
-    ]
+    layout = TWO_ARRAY_ZARR_LAYOUT[zarr_format_of(zarr_dandiset.dspath / "sample.zarr")]
     (zarr_path / ".git").touch()
     (zarr_path / ".dandi").mkdir()
     (zarr_path / ".dandi" / "somefile.txt").touch()
@@ -499,12 +496,16 @@ def test_download_different_zarr_delete_dir(
     d = new_dandiset.dandiset
     dspath = new_dandiset.dspath
     zarr.save(dspath / "sample.zarr", np.eye(5))
-    # The single-array ``np.eye(5)`` Zarr is flat (no subdirs) in V2; V3
-    # adds a ``c/`` chunk directory. The two-array Zarr below always has
-    # ``arr_0/`` and ``arr_1/`` subdirs in both V2 and V3, which is the
-    # contrast this test exercises.
+    # The single-array ``np.eye(5)`` Zarr has a flatter layout than the
+    # two-array Zarr below: in V2 no subdirs at all; in V3 only ``c/``
+    # (the chunk container) rather than ``arr_0/`` and ``arr_1/``. This
+    # is the contrast the test exercises — downloading the smaller Zarr
+    # must delete the extra ``arr_*`` subdirs of the bigger one.
+    initial_subdirs = {p.name for p in (dspath / "sample.zarr").iterdir() if p.is_dir()}
     if zarr_format_of(dspath / "sample.zarr") == "2":
-        assert not any(p.is_dir() for p in (dspath / "sample.zarr").iterdir())
+        assert initial_subdirs == set()
+    else:
+        assert initial_subdirs == {"c"}
     new_dandiset.upload()
     dd = tmp_path / d.identifier
     dd.mkdir(parents=True, exist_ok=True)
@@ -933,7 +934,7 @@ def test_download_multiple_urls(
     zarr_dandiset = sample_dandiset_factory.mkdandiset("Sample Zarr Dandiset")
     zarr_local = zarr_dandiset.dspath / "sample.zarr"
     zarr.save(zarr_local, np.arange(1000), np.arange(1000, 0, -1))
-    layout = _TWO_ARRAY_ZARR_LAYOUT[zarr_format_of(zarr_local)]
+    layout = TWO_ARRAY_ZARR_LAYOUT[zarr_format_of(zarr_local)]
     zarr_dandiset.upload()
 
     download([text_dandiset.dandiset.api_url, zarr_dandiset.dandiset.api_url], tmp_path)
